+++ services/external/email/errors.ts
